+++ src/main/java/com/onebridge/ouch/controller/user/UserController.java
@@ -1,8 +1,12 @@
 package com.onebridge.ouch.controller.user;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -33,6 +37,21 @@ public class UserController {
 	public ResponseEntity<ApiResponse<Void>> deactivateUser(
 		@UserId Long userId) {
 		userService.deactivateUser(userId);
+		return ResponseEntity.ok(ApiResponse.successWithNoData());
+	}
+
+	@PatchMapping("/languages")
+	public ResponseEntity<ApiResponse<Void>> updateUserLanguage(@AuthenticationPrincipal OuchUserDetails userDetails, @PathVariable("languageId") Long id,
+		@RequestBody String language) {
+		Long userId = userDetails.getDatabaseId();
+		userService.updateUserLanguage(userId, language);
+		return ResponseEntity.ok(ApiResponse.successWithNoData());
+	}
+
+	@GetMapping("/languages")
+	public ResponseEntity<ApiResponse<?>> getUserLanguage(@AuthenticationPrincipal OuchUserDetails userDetails) {
+		Long userId = userDetails.getDatabaseId();
+		userService.updateLanguage(userId);
 		return ResponseEntity.ok(ApiResponse.successWithNoData());
 	}
 }
