--- src/main/java/com/onebridge/ouch/dto/MessageResponse.java
@@ -1,11 +1,0 @@
-package com.onebridge.ouch.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class MessageResponse {
-
-	private String message;
-}
