--- src/main/java/com/onebridge/ouch/dto/MessageResponseDetailed.java
@@ -1,17 +1,0 @@
-package com.onebridge.ouch.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class MessageResponseDetailed<T> {
-
-	private boolean success;
-
-	private String code;
-
-	private String message;
-
-	private T data;
-}
