--- src/main/java/com/onebridge/ouch/dto/language/request/AllLangsRequest.java
@@ -1,13 +1,0 @@
-package com.onebridge.ouch.dto.language.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class AllLangsRequest {
-
-	private Long id;
-
-	private String name;
-}
