--- src/main/java/com/onebridge/ouch/dto/language/response/UserLangResponse.java
@@ -1,13 +1,0 @@
-package com.onebridge.ouch.dto.language.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserLangResponse {
-
-	private Long language_id;
-
-	private String language_name;
-}
